* #86: Applied review findings
