+++ src/main/java/com/exasol/releasedroid/github/GitHubPlatformValidator.java
@@ -5,10 +5,7 @@ import com.exasol.releasedroid.usecases.*;
 import com.exasol.releasedroid.usecases.validate.AbstractPlatformValidator;
 
 import java.util.*;
-import java.util.logging.Level;
 import java.util.logging.Logger;
-
-import static com.exasol.releasedroid.main.LoggingTool.logRedMessage;
 
 /**
  * This class checks if the project is ready for a release on GitHub.
@@ -93,7 +90,7 @@ public class GitHubPlatformValidator extends AbstractPlatformValidator {
                     + "' file before you release: " + wrongTicketsString + ".";
             report.addResult(ValidationResult
                     .successfulValidation("Skipping mentioned GitHub tickets validation. " + warningMessage));
-            logRedMessage(LOGGER, Level.WARNING, warningMessage);
+            LOGGER.warning(warningMessage);
         }
         return report;
     }
