+++ src/main/java/com/exasol/releasedroid/main/ReportLogger.java
@@ -1,26 +1,31 @@
 package com.exasol.releasedroid.main;
 
-import java.util.logging.Level;
-import java.util.logging.Logger;
-
 import com.exasol.releasedroid.usecases.Report;
 
-public class ReportLogger {
+import java.util.logging.Logger;
 
+/**
+ * This class logs report results.
+ */
+public class ReportLogger {
+    private static final String ANSI_GREEN = "\u001B[32m";
+    private static final String ANSI_RESET = "\u001B[0m";
     private static final Logger LOGGER = Logger.getLogger(ReportLogger.class.getName());
 
+    private ReportLogger() {
+        // prevent instantiation
+    }
+
     /**
-     * Log report in colors.
+     * Log report results.
      *
-     * @param logger instance of {@link Logger}
      * @param report report to log
      */
-    public static void logResults(final Logger logger, final Report report) {
-        LOGGER.addHandler(handler);
+    public static void logResults(final Report report) {
         if (report.hasFailures()) {
-            logRedMessage(logger, Level.SEVERE, report.getShortDescription() + " " + report.getFailuresReport());
+            LOGGER.severe(() -> report.getShortDescription() + " " + report.getFailuresReport());
         } else {
-            logGreenMessage(logger, Level.INFO, report.getShortDescription());
+            LOGGER.info(() -> ANSI_GREEN + report.getShortDescription() + ANSI_RESET);
         }
     }
 }
