+++ src/main/resources/logging.properties
@@ -1,6 +1,6 @@
 handlers=java.util.logging.ConsoleHandler
 .level=INFO
 java.util.logging.ConsoleHandler.level=ALL
-java.util.logging.ConsoleHandler.formatter=java.util.logging.NastyaFormatter
+java.util.logging.ConsoleHandler.formatter=com.exasol.releasedroid.main.ReleaseDroidFormatter
 java.util.logging.SimpleFormatter.format=%1$tF %1$tT.%1$tL [%4$-7s] %5$s %n
 com.exasol=ALL
